+++ src/test/java/com/project/openmarket/api/seller/SellerApiTest.java
@@ -1,4 +1,4 @@
-package com.project.openmarket.user.seller;
+package com.project.openmarket.api.seller;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import com.project.openmarket.ApiTest;
+import com.project.openmarket.api.ApiTest;
 
 class SellerApiTest extends ApiTest {
 	@Test
