+++ src/test/java/com/project/openmarket/service/ServiceTestMock.java
@@ -1,9 +1,11 @@
-package com.project.openmarket.user;
+package com.project.openmarket.service;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import com.project.openmarket.domain.product.entity.Product;
+import com.project.openmarket.domain.product.repository.ProductRepository;
 import com.project.openmarket.domain.user.entity.Consumer;
 import com.project.openmarket.domain.user.entity.Seller;
 import com.project.openmarket.domain.user.repository.ConsumerRepository;
@@ -16,7 +18,11 @@ public abstract class ServiceTestMock {
 	@Mock
 	protected SellerRepository sellerRepository;
 	@Mock
+	protected ProductRepository productRepository;
+	@Mock
 	protected Consumer consumer;
 	@Mock
 	protected Seller seller;
+	@Mock
+	protected Product product;
 }
